moved test to __testing__
