+++ packages/svg/src/__testing__/BusIcon.test.tsx
@@ -1,14 +1,14 @@
 import { render } from '@testing-library/react';
-import CircleIcon from './CircleIcon';
+import { BusIcon } from '../icons';
 
-describe('CircleIcon', () => {
+describe('BusIcon', () => {
   it('renders without errors', () => {
-    render(<CircleIcon width={24} height={24} />);
+    render(<BusIcon width={24} height={24} />);
   });
 
   it('applies width and height', () => {
-    const { getByTestId } = render(<CircleIcon width={24} height={24} />)
-    const svgElement = getByTestId('circle-icon-svg')
+    const { getByTestId } = render(<BusIcon width={24} height={24} />)
+    const svgElement = getByTestId('bus-icon-svg')
     expect(svgElement.getAttribute('width')).toBe('24');
     expect(svgElement.getAttribute('height')).toBe('24');
   })
