+++ packages/svg/src/__testing__/CircleIcon.test.tsx
@@ -1,14 +1,14 @@
 import { render } from '@testing-library/react';
-import BusIcon from './BusIcon';
+import { CircleIcon } from '../icons';
 
-describe('BusIcon', () => {
+describe('CircleIcon', () => {
   it('renders without errors', () => {
-    render(<BusIcon width={24} height={24} />);
+    render(<CircleIcon width={24} height={24} />);
   });
 
   it('applies width and height', () => {
-    const { getByTestId } = render(<BusIcon width={24} height={24} />)
-    const svgElement = getByTestId('bus-icon-svg')
+    const { getByTestId } = render(<CircleIcon width={24} height={24} />)
+    const svgElement = getByTestId('circle-icon-svg')
     expect(svgElement.getAttribute('width')).toBe('24');
     expect(svgElement.getAttribute('height')).toBe('24');
   })
